requirements.txt
# ==============================
# Web server
# ==============================

Django



# ==============================
# Database and Model Management
# ==============================

psycopg2
south
django-reversion



# ==============================
# Task Queing and Management
# ==============================

celery
django-celery



# ==============================
# Meetings and events
# ==============================
#
#git+git://github.com/openplans/meetingmatters.git#egg=meetingmatters



# ==============================
# Feeds
# ==============================

feedparser  # RSS / ATOM



# ==============================
# Authentication
# ==============================

django-social-auth



# ==============================
# APIs
# ==============================

djangorestframework



# ==============================
# Routes and Templating
# ==============================

django-jstemplate
django-shorturls



# ==============================
# Testing / QA
# ==============================

nose
django-nose
django-debug-toolbar